fix: refator date logic to separate file
