+++ crates/neon-runtime/src/napi/bindings/functions.rs
@@ -19,10 +19,6 @@ mod napi1 {
 
         fn get_value_double(env: Env, value: Value, result: *mut f64) -> Status;
 
-        fn create_date(env: Env, value: f64, result: *mut Value) -> Status;
-
-        fn get_date_value(env: Env, value: Value, result: *mut f64) -> Status;
-
         fn create_array_with_length(env: Env, length: usize, result: *mut Value) -> Status;
 
         fn get_array_length(env: Env, value: Value, result: *mut u32)-> Status;
@@ -57,7 +53,6 @@ mod napi1 {
         fn is_buffer(env: Env, value: Value, result: *mut bool) -> Status;
         fn is_error(env: Env, value: Value, result: *mut bool) -> Status;
         fn is_array(env: Env, value: Value, result: *mut bool) -> Status;
-        fn is_date(env: Env, value: Value, result: *mut bool) -> Status;
 
         fn get_value_string_utf8(
             env: Env,
@@ -256,6 +251,19 @@ mod napi4 {
     });
 }
 
+#[cfg(feature = "napi-5")]
+mod napi5 {
+    use super::super::types::*;
+
+    generate!(extern "C" {
+        fn create_date(env: Env, value: f64, result: *mut Value) -> Status;
+
+        fn get_date_value(env: Env, value: Value, result: *mut f64) -> Status;
+
+        fn is_date(env: Env, value: Value, result: *mut bool) -> Status;
+    });
+}
+
 #[cfg(feature = "napi-6")]
 mod napi6 {
     use super::super::types::*;
@@ -275,6 +283,8 @@ mod napi6 {
 pub(crate) use napi1::*;
 #[cfg(feature = "napi-4")]
 pub(crate) use napi4::*;
+#[cfg(feature = "napi-5")]
+pub(crate) use napi5::*;
 #[cfg(feature = "napi-6")]
 pub(crate) use napi6::*;
 
@@ -307,6 +317,9 @@ pub(crate) unsafe fn load(env: Env) -> Result<(), libloading::Error> {
 
     #[cfg(feature = "napi-4")]
     napi4::load(&host, version, 4)?;
+
+    #[cfg(feature = "napi-5")]
+    napi5::load(&host, version, 5)?;
 
     #[cfg(feature = "napi-6")]
     napi6::load(&host, version, 6)?;
