+++ test/napi/src/js/date.rs
@@ -2,19 +2,19 @@ use neon::prelude::*;
 use neon::types::JsDate;
 
 pub fn create_date(mut cx: FunctionContext) -> JsResult<JsDate> {
-    let date = cx.date(31415);
+    let date = JsDate::new_lossy(&mut cx, 31415);
     Ok(date)
 }
 
 pub fn create_date_from_value(mut cx: FunctionContext) -> JsResult<JsDate> {
     let time = cx.argument::<JsNumber>(0)?.value(&mut cx);
-    let date = cx.date(time);
+    let date = JsDate::new_lossy(&mut cx, time);
     Ok(date)
 }
 
 pub fn check_date_is_valid(mut cx: FunctionContext) -> JsResult<JsBoolean> {
     let time = cx.argument::<JsNumber>(0)?.value(&mut cx);
-    let date = cx.date(time);
+    let date = JsDate::new_lossy(&mut cx, time);
     let is_valid = date.is_valid(&mut cx);
     Ok(cx.boolean(is_valid))
 }
@@ -26,8 +26,8 @@ pub fn try_new_date(mut cx: FunctionContext) -> JsResult<JsUndefined> {
 }
 
 pub fn try_new_lossy_date(mut cx: FunctionContext) -> JsResult<JsUndefined> {
-    let date_overflow = JsDate::new_lossy(&mut cx, JsDate::MAX_VALUE + 1.0);
-    let date_underflow = JsDate::new_lossy(&mut cx, JsDate::MIN_VALUE - 1.0);
+    let date_overflow = JsDate::new(&mut cx, JsDate::MAX_VALUE + 1.0);
+    let date_underflow = JsDate::new(&mut cx, JsDate::MIN_VALUE - 1.0);
     assert_eq!(date_overflow.unwrap_err().kind(), neon::types::DateErrorKind::Overflow);
     assert_eq!(date_underflow.unwrap_err().kind(), neon::types::DateErrorKind::Underflow);
     Ok(cx.undefined())
@@ -35,7 +35,7 @@ pub fn try_new_lossy_date(mut cx: FunctionContext) -> JsResult<JsUndefined> {
 
 pub fn check_date_is_invalid(mut cx: FunctionContext) -> JsResult<JsBoolean> {
     let time = JsDate::MIN_VALUE - 1.0;
-    let date = cx.date(time);
+    let date = JsDate::new_lossy(&mut cx, time);
     let is_valid = date.is_valid(&mut cx);
     let val = date.value(&mut cx);
     Ok(cx.boolean(is_valid))
@@ -43,35 +43,35 @@ pub fn check_date_is_invalid(mut cx: FunctionContext) -> JsResult<JsBoolean> {
 
 pub fn create_and_get_invalid_date(mut cx: FunctionContext) -> JsResult<JsNumber> {
     let time = JsDate::MAX_VALUE + 1.0;
-    let date = cx.date(time).value(&mut cx);
-    assert!(!cx.date(time).is_valid(&mut cx));
-    assert!(cx.date(time).value(&mut cx).is_nan());
+    let date = JsDate::new_lossy(&mut cx, time).value(&mut cx);
+    assert!(!JsDate::new_lossy(&mut cx, time).is_valid(&mut cx));
+    assert!(JsDate::new_lossy(&mut cx, time).value(&mut cx).is_nan());
 
     let time = JsDate::MIN_VALUE - 1.0;
-    let date = cx.date(time).value(&mut cx);
-    assert!(!cx.date(time).is_valid(&mut cx));
-    assert!(cx.date(time).value(&mut cx).is_nan());
+    let date = JsDate::new_lossy(&mut cx, time).value(&mut cx);
+    assert!(!JsDate::new_lossy(&mut cx, time).is_valid(&mut cx));
+    assert!(JsDate::new_lossy(&mut cx, time).value(&mut cx).is_nan());
 
     let time = JsDate::MAX_VALUE + 2.0;
-    let date = cx.date(time).value(&mut cx);
-    assert!(!cx.date(time).is_valid(&mut cx));
-    assert!(cx.date(time).value(&mut cx).is_nan());
+    let date = JsDate::new_lossy(&mut cx, time).value(&mut cx);
+    assert!(!JsDate::new_lossy(&mut cx, time).is_valid(&mut cx));
+    assert!(JsDate::new_lossy(&mut cx, time).value(&mut cx).is_nan());
 
     let time = JsDate::MAX_VALUE + 3.0;
-    let date = cx.date(time).value(&mut cx);
-    assert!(!cx.date(time).is_valid(&mut cx));
-    assert!(cx.date(time).value(&mut cx).is_nan());
+    let date = JsDate::new_lossy(&mut cx, time).value(&mut cx);
+    assert!(!JsDate::new_lossy(&mut cx, time).is_valid(&mut cx));
+    assert!(JsDate::new_lossy(&mut cx, time).value(&mut cx).is_nan());
 
     let time = JsDate::MAX_VALUE + 1_000.0;
-    let date = cx.date(time).value(&mut cx);
-    assert!(!cx.date(time).is_valid(&mut cx));
-    assert!(cx.date(time).value(&mut cx).is_nan());
+    let date = JsDate::new_lossy(&mut cx, time).value(&mut cx);
+    assert!(!JsDate::new_lossy(&mut cx, time).is_valid(&mut cx));
+    assert!(JsDate::new_lossy(&mut cx, time).value(&mut cx).is_nan());
 
     Ok(cx.number(date))
 }
 
 pub fn get_date_value(mut cx: FunctionContext) -> JsResult<JsNumber> {
-    let date = cx.date(31415);
+    let date = JsDate::new_lossy(&mut cx, 31415);
     let value = date.value(&mut cx);
     println!("{:?}", value);
     Ok(cx.number(value))
